refactor: separate models from tests, organize issuance handlers
- Move issuance start tests from models to tests/issuance_start.rs
- Create src/server/handlers/issuance/ folder with mod.rs and start.rs
- Remove tests from src/domain/models/issuance/start.rs (models should just contain models)
- Rename handler to start.rs for consistency with future endpoints
