+++ src/server/handlers.rs
@@ -3,7 +3,6 @@ mod issuance;
 mod root;
 mod tenant;
 
-// Export handlers for use in the server
 pub use health::health_check;
 pub use issuance::start_issuance;
 pub use root::home;
