+++ src/server/handlers/issuance/mod.rs
@@ -1,0 +1,3 @@
+mod start;
+
+pub use start::start_issuance;
