+++ src/server/handlers/issuance/start.rs
@@ -1,9 +1,8 @@
 use axum::{Json, extract::{Extension, State}, http::StatusCode, response::IntoResponse};
 use uuid::Uuid;
 
-use crate::domain::models::issuance::{
-    start_issuance_session, StartIssuanceRequest,
-};
+use crate::domain::models::issuance::start_issuance_session;
+use crate::domain::models::issuance::StartIssuanceRequest;
 use crate::server::error::IntoApiError;
 use crate::server::{AppState, responses::ResponseBody};
 use crate::session::SessionStore;
